+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Allow users to manage the downloading of WASM grammars not bundled with treetags
 - Add extension fields support for OCaml
 - Add extension fields support for PHP
 - Add extension fields support for Ruby
+- Add extension fields support for Bash
 
 ### Changed
 - Tag Address is truncated to 96 chars for native parsers